isomorphic-flux-starter spinoff: First pass
This is basically a super dumbed down, barebones version of capacitor.
Capacitor is far superior, but introduces a lot of concepts at once.
This spinoff tries to limit the scope in order to make an isomorphic
flux setup as approachable as possible.

The name and structure changes are just personal preference. I just
organized things in a way that felt digestible to someone who's never
worked with these tools.

This should eventually get moved to it's own repo, as it's purpose is
a little different than capacitor. I think we should use Capacitor to
keep up with more advanced combinations of toolsets and techniques,
while keeping isomorphic-flux-starter entry level and bare-bones.