Setup Hapi
This comment replaces Express with Hapi for the main application
server. It sets up a proxy to webpack-dev-server so that assets have
access to features such as hot module replacement.

Additionally, I changed the way actions are accessed, following some
interview material with Facebook staff working on their flux apps.